Separate business logic from KV store interface
register() and writeTerminalStatus() are orchestration — cross-store
writes, cache invalidation, error-swallowing policy. They don't belong
on a storage interface. Move them to executionLifecycle.ts as free
functions that operate on the store, breaking the circular dependency
between ExecutionKVStore and executionListing.

Dependency graph is now a clean DAG:
  executionLifecycle → ExecutionKVStore, executionListing
  executionListing   → ExecutionKVStore
