+++ src/tools/bash.ts
@@ -14,7 +14,11 @@ import {
   untrackExecution,
   ProcessExecutionHandle,
 } from '@agent/runtime/executionRegistry';
-import { getExecutionStore } from '@agent/storage';
+import {
+  getExecutionStore,
+  registerExecution,
+  writeTerminalStatus,
+} from '@agent/storage';
 import { AgentConfigSchema } from '@agent/core/AgentConfig';
 import { ToolUseFollowUpQueue } from '@agent/toolUse/ToolUseFollowUpQueueManager';
 
@@ -206,9 +210,9 @@ export class BashTool extends defineTool({
     );
     handle.outputPaths = { stdout: stdoutPath, stderr: stderrPath };
 
-    const store = getExecutionStore(executionId);
     try {
-      await store.register(
+      await registerExecution(
+        executionId,
         syntheticConfig,
         'bash',
         parentExecutionId,
@@ -254,7 +258,7 @@ export class BashTool extends defineTool({
         ]);
 
         const terminalStatus = result.success ? 'completed' : 'error';
-        await store.writeTerminalStatus(terminalStatus).catch(() => {});
+        await writeTerminalStatus(executionId, terminalStatus).catch(() => {});
         untrackExecution(executionId);
 
         const msg = formatBashDelivery(
@@ -269,11 +273,11 @@ export class BashTool extends defineTool({
         ToolUseFollowUpQueue.enqueue(parentStreamId, msg);
       })
       .catch(async (err: unknown) => {
-        await store.writeTerminalStatus('error').catch(() => {});
+        await writeTerminalStatus(executionId, 'error').catch(() => {});
         untrackExecution(executionId);
 
         const msg = formatBashError(executionId, command, err);
-        await store.write('report', msg);
+        await getExecutionStore(executionId).write('report', msg);
         ToolUseFollowUpQueue.enqueue(parentStreamId, msg);
       })
       .finally(cleanupTempFiles);
